refactor: move LinkAccounts to separate file
