+++ src/components/layout/dialogs/UserProfile.jsx
@@ -1,13 +1,10 @@
 import * as React from 'react'
 import {
   Grid,
-  Typography,
   DialogContent,
   AppBar,
   Tabs,
   Tab,
-  Select,
-  MenuItem,
   TextField,
   Box,
 } from '@mui/material'
@@ -21,13 +18,11 @@ import Query from '@services/Query'
 import Header from '../general/Header'
 import Footer from '../general/Footer'
 import TabPanel from '../general/TabPanel'
-import DiscordLogin from '../auth/Discord'
-import Telegram from '../auth/Telegram'
-import Notification from '../general/Notification'
 import { DialogWrapper } from './DialogWrapper'
 import { UserBackups } from './profile/Backups'
 import { UserPermissions } from './profile/Permissions'
 import { UserGymBadges } from './profile/GymBadges'
+import { LinkAccounts } from './profile/LinkAccounts'
 
 export default function UserProfile() {
   Utility.analytics('/user-profile')
@@ -74,7 +69,7 @@ export default function UserProfile() {
           minHeight="70vh"
         >
           <TabPanel value={tab} index={0}>
-            <LinkProfiles />
+            <LinkAccounts />
             <ExtraFields />
             <UserBackups />
           </TabPanel>
@@ -106,98 +101,6 @@ export default function UserProfile() {
         ]}
       />
     </DialogWrapper>
-  )
-}
-
-const LinkProfiles = () => {
-  const { t } = useTranslation()
-  const auth = useStatic((s) => s.auth)
-  const { discordAuthUrl, telegramAuthUrl, telegramBotName } = useStatic(
-    (state) => state.config.customRoutes,
-  )
-
-  const [refreshing, setRefreshing] = React.useState(false)
-
-  const [setWebhookStrategy] = useMutation(Query.user('setWebhookStrategy'))
-
-  if (refreshing) {
-    setTimeout(() => window.location.reload(), 2000)
-  }
-
-  return (
-    <>
-      <Grid container alignItems="center" justifyContent="center">
-        {['discord', 'telegram'].map((method, i) => {
-          if (!auth.methods.includes(method)) return null
-          const Component = i ? (
-            <Telegram authUrl={telegramAuthUrl} botName={telegramBotName} />
-          ) : (
-            <DiscordLogin
-              href={discordAuthUrl}
-              text="link_discord"
-              size="medium"
-            />
-          )
-          return (
-            <Grid item xs={12} key={method} align="center">
-              {auth[`${method}Id`] ? (
-                <Typography color="secondary">
-                  {t('user_username')}: {auth.username} ({t(`${method}_linked`)}
-                  )
-                </Typography>
-              ) : (
-                Component
-              )}
-            </Grid>
-          )
-        })}
-        {auth.discordId && auth.telegramId && (
-          <Grid container item alignItems="center" justifyContent="center">
-            <Grid
-              item
-              xs={6}
-              sm={6}
-              md={5}
-              style={{ textAlign: 'center', padding: '20px 0' }}
-            >
-              <Typography>{t('select_webhook_strategy')}</Typography>
-            </Grid>
-            <Grid item xs={6} sm={6} md={5} style={{ textAlign: 'center' }}>
-              <Select
-                value={auth.webhookStrategy || ''}
-                onChange={(e) => {
-                  setWebhookStrategy({
-                    variables: {
-                      strategy: e.target.value,
-                    },
-                  })
-                  useStatic.setState({
-                    auth: { ...auth, webhookStrategy: e.target.value },
-                  })
-                  setRefreshing(true)
-                }}
-                style={{ minWidth: 100 }}
-              >
-                {['discord', 'telegram'].map((x) =>
-                  auth[`${x}Id`] ? (
-                    <MenuItem key={x} value={x}>
-                      {Utility.getProperName(x)}
-                    </MenuItem>
-                  ) : null,
-                )}
-              </Select>
-            </Grid>
-          </Grid>
-        )}
-      </Grid>
-      {refreshing && (
-        <Notification
-          severity="success"
-          i18nKey="webhook_strategy_success"
-          messages={[{ key: 'redirecting', variables: [] }]}
-        />
-      )}
-    </>
   )
 }
